// src/app/api/models/customer-payment.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { Allocation } from '../models/allocation';
import { CurrencyEnum } from '../models/currency-enum';
import { PaymentStatusEnum } from '../models/payment-status-enum';
import { PaymentTenderEnum } from '../models/payment-tender-enum';
export interface CustomerPayment {
  allocations?: Array<Allocation>;

  /**
   * In US dollars.
   */
  amount: string;
  amount_tendered: string;
  currency?: CurrencyEnum;
  customer: number;
  customer_name: string;
  exchange_rate: string | null;
  id: number;
  note?: string;
  number: string;
  payment_date?: string;
  reference?: string;
  status: PaymentStatusEnum;
  taken_by_name: string | null;
  tender: PaymentTenderEnum;
  void_reason: string;
  voided_at: string | null;
  voided_by_name: string | null;
}